added a base template and moved some common methods into it
