+++ src/main/groovy/com/toastcoders/vmware/yavijava/data/BaseTemplate.groovy
@@ -1,0 +1,50 @@
+package com.toastcoders.vmware.yavijava.data;
+
+/**
+ * Copyright 2015 Michael Rice <michael@michaelrice.org>
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+public class BaseTemplate {
+
+    public static String getPackage() {
+        return "package com.vmware.vim25;\n"
+    }
+
+    public static String getLicense() {
+        Date today = new Date()
+        return """
+/**
+ * Created by Michael Rice on ${today}
+ *
+ * Copyright 2015 Michael Rice
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * @since 6.0
+ */\n\n"""
+    }
+
+    public static String closeClass() {
+        return "}\n"
+    }
+}
